requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
Jinja2==3.1.3
python-multipart==0.0.9
email-validator==2.1.1
# Testing
pytest==8.1.1
passlib==1.7.4
coverage==7.4.4
# Database
SQLAlchemy==2.0.38
# LLM
langchain-community==0.3.9
langchain==0.3.9
langchain-experimental==0.3.4
tiktoken==0.8.0
langchain-openai==0.2.10
langgraph==0.2.60
# sentence_transformers==2.3.1
psycopg2==2.9.10
pandas==2.2.3
openpyxl==3.1.5
bcrypt==4.2.1
# Tabular
tabulate==0.9.0
XlsxWriter==3.2.2
#Authentication
pyjwt==2.10.1